// secret-share-client/generated/models/ICreateSecretRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ICreateSecretRequest = {
    encryptedSecret?: string;
    ipRestrictions?: Array<string>;
    maxViews?: number;
    secretPassword?: string;
    expirationDate?: string;
    emailNotification?: string;
};